requirements.txt
# Core ML and Deep Learning
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
scikit-learn==1.3.0

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
requests==2.31.0